[refactor] separate category interface and service
